+++ backend/app/api/routes/users.py
@@ -4,7 +4,7 @@ from typing import Any
 from fastapi import APIRouter, Depends, HTTPException
 from sqlmodel import col, delete, func, select
 
-from app import crud
+import app.crud.user as crud
 from app.api.deps import (
     CurrentUser,
     SessionDep,
