+++ backend/app/crud/inquiry.py
@@ -1,0 +1,19 @@
+from sqlmodel import Session, select
+
+from app.models import (
+    Inquiry,
+    InquiryCreate,
+)
+
+def create_inquiry(*, session: Session, inquiry_in: InquiryCreate) -> Inquiry:
+    db_inquiry = Inquiry.model_validate(inquiry_in)
+    session.add(db_inquiry)
+    session.commit()
+    session.refresh(db_inquiry)
+    return db_inquiry
+
+
+def get_inquiry_by_text(*, session: Session, text: str) -> Inquiry | None:
+    statement = select(Inquiry).where(Inquiry.text == text)
+    session_text = session.exec(statement).first()
+    return session_text
