+++ backend/app/crud/item.py
@@ -1,0 +1,18 @@
+import uuid
+
+from sqlmodel import Session, select
+
+from app.models import (
+    Item,
+    ItemCreate,
+)
+
+
+
+def create_item(*, session: Session, item_in: ItemCreate, owner_id: uuid.UUID) -> Item:
+    db_item = Item.model_validate(item_in, update={"owner_id": owner_id})
+    session.add(db_item)
+    session.commit()
+    session.refresh(db_item)
+    return db_item
+
