+++ backend/app/crud/user.py
@@ -1,19 +1,13 @@
-import uuid
 from typing import Any
 
 from sqlmodel import Session, select
 
 from app.core.security import get_password_hash, verify_password
 from app.models import (
-    Inquiry,
-    InquiryCreate,
-    Item,
-    ItemCreate,
     User,
     UserCreate,
     UserUpdate,
 )
-
 
 def create_user(*, session: Session, user_create: UserCreate) -> User:
     db_obj = User.model_validate(
@@ -53,24 +47,3 @@ def authenticate(*, session: Session, email: str, password: str) -> User | None:
         return None
     return db_user
 
-
-def create_item(*, session: Session, item_in: ItemCreate, owner_id: uuid.UUID) -> Item:
-    db_item = Item.model_validate(item_in, update={"owner_id": owner_id})
-    session.add(db_item)
-    session.commit()
-    session.refresh(db_item)
-    return db_item
-
-
-def create_inquiry(*, session: Session, inquiry_in: InquiryCreate) -> Inquiry:
-    db_inquiry = Inquiry.model_validate(inquiry_in)
-    session.add(db_inquiry)
-    session.commit()
-    session.refresh(db_inquiry)
-    return db_inquiry
-
-
-def get_inquiry_by_text(*, session: Session, text: str) -> Inquiry | None:
-    statement = select(Inquiry).where(Inquiry.text == text)
-    session_text = session.exec(statement).first()
-    return session_text
