+++ backend/app/tests/api/routes/test_users.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 from fastapi.testclient import TestClient
 from sqlmodel import Session, select
 
-from app import crud
+import app.crud.user as crud
 from app.core.config import settings
 from app.core.security import verify_password
 from app.models import User, UserCreate
