+++ backend/app/tests/crud/test_user.py
@@ -1,7 +1,7 @@
 from fastapi.encoders import jsonable_encoder
 from sqlmodel import Session
 
-from app import crud
+import app.crud.user as crud
 from app.core.security import verify_password
 from app.models import User, UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
