+++ backend/app/tests/utils/user.py
@@ -1,7 +1,7 @@
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
-from app import crud
+import app.crud.user as crud
 from app.core.config import settings
 from app.models import User, UserCreate, UserUpdate
 from app.tests.utils.utils import random_email, random_lower_string
